@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 
 class Tile extends StatelessWidget {
   const Tile({
-    Key key,
+    Key? key,
     this.color = const Color(0xFFF4F4F8),
-    @required this.text,
+    required this.text,
   }) : super(key: key);
 
   final Color color;
