@@ -5,7 +5,7 @@ void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
   const MyApp({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -26,7 +26,7 @@ class MyApp extends StatelessWidget {
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -38,7 +38,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) {
-    final direction = AppState.of(context).direction;
+    final direction = AppState.of(context)!.direction;
     return Scaffold(
       appBar: AppBar(
         title: const Text('Flutter Slidable'),
@@ -123,29 +123,30 @@ class _MyHomePageState extends State<MyHomePage> {
                     },
                     closeOnCancel: true,
                     confirmDismiss: () async {
-                      return showDialog<bool>(
-                        context: context,
-                        builder: (context) {
-                          return AlertDialog(
-                            title: const Text('Are you sure?'),
-                            content: const Text('Are you sure to dismiss?'),
-                            actions: [
-                              TextButton(
-                                onPressed: () {
-                                  Navigator.of(context).pop(true);
-                                },
-                                child: const Text('Yes'),
-                              ),
-                              TextButton(
-                                onPressed: () {
-                                  Navigator.of(context).pop(false);
-                                },
-                                child: const Text('No'),
-                              ),
-                            ],
-                          );
-                        },
-                      );
+                      return await showDialog<bool>(
+                            context: context,
+                            builder: (context) {
+                              return AlertDialog(
+                                title: const Text('Are you sure?'),
+                                content: const Text('Are you sure to dismiss?'),
+                                actions: [
+                                  TextButton(
+                                    onPressed: () {
+                                      Navigator.of(context).pop(true);
+                                    },
+                                    child: const Text('Yes'),
+                                  ),
+                                  TextButton(
+                                    onPressed: () {
+                                      Navigator.of(context).pop(false);
+                                    },
+                                    child: const Text('No'),
+                                  ),
+                                ],
+                              );
+                            },
+                          ) ??
+                          false;
                     },
                   ),
                   children: const [
@@ -243,9 +244,9 @@ class _MyHomePageState extends State<MyHomePage> {
 
 class SlideAction extends StatelessWidget {
   const SlideAction({
-    Key key,
-    @required this.color,
-    @required this.icon,
+    Key? key,
+    required this.color,
+    required this.icon,
     this.flex = 1,
   }) : super(key: key);
 
@@ -268,9 +269,9 @@ class SlideAction extends StatelessWidget {
 
 class Tile extends StatelessWidget {
   const Tile({
-    Key key,
-    @required this.color,
-    @required this.text,
+    Key? key,
+    required this.color,
+    required this.text,
   }) : super(key: key);
 
   final Color color;
@@ -278,10 +279,10 @@ class Tile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final direction = AppState.of(context).direction;
+    final direction = AppState.of(context)!.direction;
     return ActionTypeListener(
       child: GestureDetector(
-        onLongPress: () => Slidable.of(context).openEndActionPane(),
+        onLongPress: () => Slidable.of(context)!.openEndActionPane(),
         child: Container(
           color: color,
           height: direction == Axis.horizontal ? 100 : double.infinity,
@@ -295,8 +296,8 @@ class Tile extends StatelessWidget {
 
 class ActionTypeListener extends StatefulWidget {
   const ActionTypeListener({
-    Key key,
-    @required this.child,
+    Key? key,
+    required this.child,
   }) : super(key: key);
 
   final Widget child;
@@ -306,7 +307,7 @@ class ActionTypeListener extends StatefulWidget {
 }
 
 class _ActionTypeListenerState extends State<ActionTypeListener> {
-  ValueNotifier<ActionPaneType> _actionPaneTypeValueNotifier;
+  ValueNotifier<ActionPaneType>? _actionPaneTypeValueNotifier;
 
   @override
   void initState() {
@@ -333,9 +334,9 @@ class _ActionTypeListenerState extends State<ActionTypeListener> {
 
 class AppState extends InheritedWidget {
   const AppState({
-    Key key,
-    @required this.direction,
-    @required Widget child,
+    Key? key,
+    required this.direction,
+    required Widget child,
   }) : super(key: key, child: child);
 
   final Axis direction;
@@ -345,7 +346,7 @@ class AppState extends InheritedWidget {
     return direction != oldWidget.direction;
   }
 
-  static AppState of(BuildContext context) {
+  static AppState? of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<AppState>();
   }
 }
