@@ -36,7 +36,7 @@ class MyApp extends StatelessWidget {
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -45,7 +45,7 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage>
     with SingleTickerProviderStateMixin {
-  AnimationController controller;
+  AnimationController? controller;
 
   @override
   void initState() {
@@ -77,10 +77,10 @@ class _MyHomePageState extends State<MyHomePage>
       ),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
-          if (controller.isCompleted) {
-            controller.reverse();
-          } else if (controller.isDismissed) {
-            controller.forward();
+          if (controller!.isCompleted) {
+            controller!.reverse();
+          } else if (controller!.isDismissed) {
+            controller!.forward();
           }
         },
         child: const Icon(Icons.play_arrow),
@@ -91,41 +91,41 @@ class _MyHomePageState extends State<MyHomePage>
 
 class SlidablePlayer extends StatefulWidget {
   const SlidablePlayer({
-    Key key,
-    @required this.animation,
-    @required this.child,
+    Key? key,
+    required this.animation,
+    required this.child,
   }) : super(key: key);
 
-  final Animation<double> animation;
+  final Animation<double>? animation;
   final Widget child;
 
   @override
   _SlidablePlayerState createState() => _SlidablePlayerState();
 
-  static _SlidablePlayerState of(BuildContext context) {
+  static _SlidablePlayerState? of(BuildContext context) {
     return context.findAncestorStateOfType<_SlidablePlayerState>();
   }
 }
 
 class _SlidablePlayerState extends State<SlidablePlayer> {
-  final Set<SlidableController> controllers = <SlidableController>{};
+  final Set<SlidableController?> controllers = <SlidableController?>{};
 
   @override
   void initState() {
     super.initState();
-    widget.animation.addListener(handleAnimationChanged);
+    widget.animation!.addListener(handleAnimationChanged);
   }
 
   @override
   void dispose() {
-    widget.animation.removeListener(handleAnimationChanged);
+    widget.animation!.removeListener(handleAnimationChanged);
     super.dispose();
   }
 
   void handleAnimationChanged() {
-    final value = widget.animation.value;
+    final value = widget.animation!.value;
     controllers.forEach((controller) {
-      controller.ratio = value;
+      controller!.ratio = value;
     });
   }
 
@@ -137,11 +137,11 @@ class _SlidablePlayerState extends State<SlidablePlayer> {
 
 class SlidableControllerSender extends StatefulWidget {
   const SlidableControllerSender({
-    Key key,
+    Key? key,
     this.child,
   }) : super(key: key);
 
-  final Widget child;
+  final Widget? child;
 
   @override
   _SlidableControllerSenderState createState() =>
@@ -149,33 +149,33 @@ class SlidableControllerSender extends StatefulWidget {
 }
 
 class _SlidableControllerSenderState extends State<SlidableControllerSender> {
-  SlidableController controller;
-  _SlidablePlayerState playerState;
+  SlidableController? controller;
+  _SlidablePlayerState? playerState;
 
   @override
   void initState() {
     super.initState();
     controller = Slidable.of(context);
     playerState = SlidablePlayer.of(context);
-    playerState.controllers.add(controller);
+    playerState!.controllers.add(controller);
   }
 
   @override
   void dispose() {
-    playerState.controllers.remove(controller);
+    playerState!.controllers.remove(controller);
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return widget.child;
+    return widget.child!;
   }
 }
 
 class MySlidable extends StatelessWidget {
   const MySlidable({
-    Key key,
-    @required this.motion,
+    Key? key,
+    required this.motion,
   }) : super(key: key);
 
   final Widget motion;
@@ -199,10 +199,10 @@ class MySlidable extends StatelessWidget {
 
 class SlideAction extends StatelessWidget {
   const SlideAction({
-    Key key,
-    @required this.color,
-    @required this.icon,
-    @required this.label,
+    Key? key,
+    required this.color,
+    required this.icon,
+    required this.label,
     this.flex = 1,
   }) : super(key: key);
 
